pyproject.toml
[tool.poetry]
name = "library-reservation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "library_reservation"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.3"
pytest = "^8.3.5"
selenium = "^4.29.0"
line-bot-sdk = "^3.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
